+++ packages/general/src/environment/Environment.ts
@@ -4,6 +4,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
+import { Boot } from "#util/Boot.js";
 import { MaybePromise } from "#util/Promises.js";
 import { DiagnosticSource } from "../log/DiagnosticSource.js";
 import { Logger } from "../log/Logger.js";
@@ -237,4 +238,8 @@ export class Environment {
     }
 }
 
-let global: Environment = new Environment("default");
+let global: Environment;
+
+Boot.init(() => {
+    global = new Environment("default");
+});