+++ packages/react-native/src/crypto/register.ts
@@ -4,13 +4,15 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import { Crypto, NoProviderError, singleton } from "@matter/general";
+import { Boot, Crypto, NoProviderError, singleton } from "@matter/general";
 import { CryptoReactNative } from "./ReactNativeCrypto.js";
 
-// Check if Crypto singleton is already registered and auto register if not
-try {
-    Crypto.get();
-} catch (error) {
-    NoProviderError.accept(error);
-    Crypto.get = singleton(() => new CryptoReactNative());
-}
+Boot.init(() => {
+    // Check if Crypto singleton is already registered and auto register if not
+    try {
+        Crypto.get();
+    } catch (error) {
+        NoProviderError.accept(error);
+        Crypto.get = singleton(() => new CryptoReactNative());
+    }
+});